bit -> calc
