+++ Makefile
@@ -21,7 +21,6 @@ OBJECTS = \
 	mnemonic_calc.o \
 	mnemonic_ex.o \
 	mnemonic_stack.o \
-	mnemonic_bit.o \
 	mnemonic_vgs.o \
 	numeric.o \
 	operand.o \
@@ -103,9 +102,6 @@ mnemonic_ex.o: src/mnemonic_ex.cpp ${HEADERS} src/mnemonic.h
 	${CPP} ${CPPFLAGS} -c $< -o $@
 
 mnemonic_stack.o: src/mnemonic_stack.cpp ${HEADERS} src/mnemonic.h
-	${CPP} ${CPPFLAGS} -c $< -o $@
-
-mnemonic_bit.o: src/mnemonic_bit.cpp ${HEADERS} src/mnemonic.h
 	${CPP} ${CPPFLAGS} -c $< -o $@
 
 mnemonic_vgs.o: src/mnemonic_vgs.cpp ${HEADERS} src/mnemonic.h
